+++ app/views/signs/index.html.erb
@@ -4,7 +4,7 @@
     <h1>My signs</h1>
     <form class="cell auto">
       <%= label_tag :sort_by %>
-      <%= select_tag :sort_by, options_for_select(Sign::SORT_OPTIONS.keys.collect{ |name| [name.to_s.humanize, name] }, params[:sort_by]),
+      <%= select_tag :sort_by, options_for_select(SignsController::SORT_OPTIONS.keys.collect{ |name| [name.to_s.humanize, name] }, params[:sort_by]),
                                onchange: "this.form.submit()" %>
     </form>
   </div>
